fix: sanitize API errors and share frontend error formatting
